ragel.rl: mark anything as error
The word label is mainly meant to define printable and conceated
characters. In print_mark we use it to print tail in case we hit the
valid* kleene star expression. Otherwise we will just show '^' mark on
whatever byte we failed on. This allows to catch semantical failures
such as having newlines at wrong place etc..